refactor(frameworks): move framework-detector + catalog from ingestion
Moves the 6 framework-detection source files out of
packages/ingestion/src/pipeline/profile-detectors/ into the new
packages/frameworks/src/ package per T-M4-7. All moves use git mv so
git blame follows the files.

Files moved:
- framework-detector.ts -> detector.ts
- frameworks-catalog.ts -> catalog.ts
- frameworks.ts -> frameworks.ts
- manifests.ts -> manifests.ts
- variant-detectors.ts -> variant-detectors.ts
- framework-detector.test.ts -> detector.test.ts

Updates:
- packages/frameworks/src/index.ts re-exports the public surface
- packages/ingestion/src/pipeline/phases/profile.ts imports from
  @opencodehub/frameworks
- packages/ingestion/package.json adds the workspace dep
- packages/ingestion/tsconfig.json adds a project reference

Cross-package type leak: frameworks.ts and manifests.ts previously
depended on ScannedFile from the ingestion scan phase. Introduced a
minimal FrameworkFileInput { relPath: string } interface so the
frameworks package has no back-reference to ingestion.
