+++ packages/frameworks/src/variant-detectors.ts
